refactor(web-app): tunings page folder structure
